+++ web/wiki/src/__tests__/organisms/ResourceForm.test.tsx
@@ -3,6 +3,8 @@ import { useLocation } from 'react-router-dom'
 import { fireEvent, render, screen, waitFor } from '../test-utils'
 import { ResourceForm } from '../../components/organisms'
 import { reloadPage } from '../../utils/navigation'
+import { errorHandlers } from '../../__mocks__/handlers'
+import { server } from '../../__mocks__/server'
 
 const reload = vi.fn(() => reloadPage)
 
@@ -15,6 +17,7 @@ vi.mock('../utils/navigation', async () => {
     reloadPage: reload,
   }
 })
+
 beforeEach(() => {
   vi.mock('react-router-dom', async () => {
     const actual: Record<number, unknown> = await vi.importActual(
@@ -55,6 +58,17 @@ const options = [
   },
 ]
 
+const initialValues = {
+  id: 'resource-id',
+  title: 'Initial Title',
+  description: 'Initial Description',
+  url: 'https://example.com',
+  topicId: 'cli04v2l0000008mq5pwx7w5j',
+  resourceType: 'VIDEO',
+}
+
+const resourceId = 'resource-id'
+
 describe('ResourceForm', () => {
   it('renders correctly', () => {
     vi.mocked(useLocation).mockReturnValue({
@@ -74,6 +88,7 @@ describe('ResourceForm', () => {
 
     expect(screen.getByText(/crear/i)).toBeInTheDocument()
   })
+
   it('renders correctly on edit resource with initial values', () => {
     vi.mocked(useLocation).mockReturnValue({
       pathname: '',
@@ -82,23 +97,18 @@ describe('ResourceForm', () => {
       key: '',
       state: { name: 'React', id: 'cln1er1vn000008mk79bs02c5' },
     }) as unknown as Location
-    const initialValues = {
-      title: 'TEST TITLE',
-      description: 'TEST DESCRIPTION',
-      url: 'TEST URL',
-      resourceType: 'VIDEO',
-      topicId: 'cli04v2l0000008mq5pwx7w5j',
-    }
+
     render(
       <ResourceForm selectOptions={options} initialValues={initialValues} />
     )
 
-    expect(screen.getByDisplayValue('TEST TITLE')).toBeInTheDocument()
-    expect(screen.getByDisplayValue('TEST DESCRIPTION')).toBeInTheDocument()
-    expect(screen.getByDisplayValue('TEST URL')).toBeInTheDocument()
+    expect(screen.getByDisplayValue('Initial Title')).toBeInTheDocument()
+    expect(screen.getByDisplayValue('Initial Description')).toBeInTheDocument()
+    expect(screen.getByDisplayValue('https://example.com')).toBeInTheDocument()
     expect(screen.getByDisplayValue('VIDEO')).toBeInTheDocument()
     expect(screen.getByDisplayValue('Listas')).toBeInTheDocument()
   })
+
   it('should show error message when form input is invalid', async () => {
     render(<ResourceForm selectOptions={options} />)
 
@@ -141,7 +151,7 @@ describe('ResourceForm', () => {
       target: { value: 'cli04uxud000609k37w9phejw' },
     })
 
-    const videoRadio = screen.getByTestId('resourceType') as HTMLInputElement
+    const videoRadio = screen.getByLabelText('Video') as HTMLInputElement
     fireEvent.click(videoRadio)
 
     const button = screen.getByText(/crear/i)
@@ -152,15 +162,95 @@ describe('ResourceForm', () => {
       expect(descriptionInput.value).toBe('TESTING DESCRIPTION')
       expect(urlInput.value).toBe('https://dev.itawiki.eurecatacademy.org/')
       expect(temaSelect.value).toBe('cli04uxud000609k37w9phejw')
-      expect(videoRadio).toBeEnabled()
+      expect(videoRadio).toBeChecked()
+
       expect(button).toBeEnabled()
     })
-    reloadPage()
+
+    await waitFor(() => {
+      expect(screen.getByTestId('done-icon')).toBeInTheDocument()
+    })
   })
 
-  // TODO: Fix when error case onError is definedw ithin the ResourceForm component
-  it.skip('should show an error if the POST request fails', async () => {
+  it('should show an error if the POST request fails', async () => {
+    vi.mocked(useLocation).mockReturnValue({
+      pathname: '',
+      search: '',
+      hash: '',
+      key: '',
+      state: { name: 'React', id: 'cln1er1vn000008mk79bs02c5' },
+    }) as unknown as Location
+
+    server.use(...errorHandlers)
+
     render(<ResourceForm selectOptions={options} />)
+
+    fireEvent.change(screen.getByLabelText(/títol/i), {
+      target: { value: 'Title' },
+    })
+    fireEvent.change(screen.getByLabelText(/descripció/i), {
+      target: { value: 'Description text' },
+    })
+    fireEvent.change(screen.getByLabelText(/url/i), {
+      target: { value: 'https://repeatedUrl/' },
+    })
+    const temaSelect = screen.getByTestId('resourceTopic') as HTMLSelectElement
+    fireEvent.change(temaSelect, {
+      target: { value: 'cli04uxud000609k37w9phejw' },
+    })
+    const blogRadio = screen.getByLabelText('Blog') as HTMLInputElement
+    fireEvent.click(blogRadio)
+
+    fireEvent.click(screen.getByRole('button', { name: 'Crear' }))
+
+    await waitFor(() =>
+      expect(screen.getByTestId('error-message')).toBeInTheDocument()
+    )
+    expect(
+      screen.getByText(
+        'Aquest recurs ja existeix. Si us plau, introdueix un recurs diferent.'
+      )
+    ).toBeInTheDocument()
+
+    expect(screen.queryByTestId('done-icon')).not.toBeInTheDocument()
+  })
+
+  it('should show an error if the PATCH request fails', async () => {
+    vi.mocked(useLocation).mockReturnValue({
+      pathname: '',
+      search: '',
+      hash: '',
+      key: '',
+      state: { name: 'React', id: 'cln1er1vn000008mk79bs02c5' },
+    }) as unknown as Location
+
+    server.use(...errorHandlers)
+
+    render(
+      <ResourceForm
+        selectOptions={options}
+        initialValues={initialValues}
+        resourceId={resourceId}
+      />
+    )
+
+    fireEvent.change(screen.getByLabelText(/títol/i), {
+      target: { value: 'Edited Title' },
+    })
+
+    fireEvent.click(screen.getByRole('button', { name: 'Edita' }))
+
+    await waitFor(() =>
+      expect(screen.getByTestId('error-message')).toBeInTheDocument()
+    )
+
+    expect(
+      screen.getByText(
+        'Error en la base de dades. Per favor, intenta-ho més tard.'
+      )
+    ).toBeInTheDocument()
+
+    expect(screen.queryByTestId('done-icon')).not.toBeInTheDocument()
   })
 
   it('should submit the form for updating a resource when initialValues is provided', async () => {
@@ -171,15 +261,6 @@ describe('ResourceForm', () => {
       key: '',
       state: { name: 'React', id: 'cln1er1vn000008mk79bs02c5' },
     }) as unknown as Location
-    const initialValues = {
-      id: 'resource-id',
-      title: 'Initial Title',
-      description: 'Initial Description',
-      url: 'https://example.com',
-      topicId: 'cli04v2l0000008mq5pwx7w5j',
-      resourceType: 'VIDEO',
-    }
-    const resourceId = 'resource-id'
 
     render(
       <ResourceForm
@@ -206,7 +287,10 @@ describe('ResourceForm', () => {
 
     const button = screen.getByTestId('submit-button')
     fireEvent.click(button)
-    reloadPage()
+
+    await waitFor(() => {
+      expect(screen.getByTestId('done-icon')).toBeInTheDocument()
+    })
   })
 
   it('should show error message when updating a resource with invalid input', async () => {
@@ -217,15 +301,6 @@ describe('ResourceForm', () => {
       key: '',
       state: { name: 'React', id: 'cln1er1vn000008mk79bs02c5' },
     }) as unknown as Location
-    const initialValues = {
-      id: 'resource-id',
-      title: 'Initial Title',
-      description: 'Initial Description',
-      url: 'https://example.com',
-      topicId: 'cli04v2l0000008mq5pwx7w5j',
-      resourceType: 'VIDEO',
-    }
-    const resourceId = 'resource-id'
 
     render(
       <ResourceForm
@@ -238,12 +313,12 @@ describe('ResourceForm', () => {
     fireEvent.change(screen.getByLabelText(/títol/i), {
       target: { value: '' },
     })
+
     const button = screen.getByTestId('submit-button')
     fireEvent.click(button)
 
     await waitFor(() =>
       expect(screen.getByText('Aquest camp és obligatori')).toBeInTheDocument()
     )
-    reloadPage()
   })
 })